requirements.txt
# Speech Understanding — Programming Assignment 2
# IIT Jodhpur | MTech Data Engineering | M25DE1047
#
# ✅ macOS 12+ Apple Silicon (M1/M2/M3/M4) — Python 3.12
#
# Install everything at once:
#   ./install.sh
#
# Or manually:
#   Step 1 — system deps (macOS):
#     brew install ffmpeg espeak-ng
#
#   Step 2 — python packages:
#     pip install -r requirements.txt

# ─── Core ────────────────────────────────────────────────────────
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
scipy>=1.10.0

# ─── ASR — Part 1 ────────────────────────────────────────────────
openai-whisper>=20231117
transformers>=4.36.0
accelerate>=0.25.0
huggingface-hub>=0.20.0

# ─── Audio processing ─────────────────────────────────────────────
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# ─── Denoising — Part 1 Task 1.3 (OPTIONAL) ──────────────────────
# Needs Rust compiler on macOS. Pipeline auto-falls-back to
# spectral subtraction if not installed.
#   Enable: brew install rust && pip install deepfilternet
# deepfilternet>=0.5.0

# ─── TTS — Part 3 ────────────────────────────────────────────────
# coqui-tts = community fork of old "TTS" package (abandoned, no py3.12)
# Supports Python 3.9–3.12, macOS ARM64 native wheels.
coqui-tts>=0.24.0

# ─── Evaluation & utilities ───────────────────────────────────────
jiwer>=3.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pandas>=2.0.0
pyyaml>=6.0
einops>=0.7.0
